+++ src/common/components/Chip/ChipColorFilled.jsx
@@ -3,8 +3,10 @@ import PropTypes from 'prop-types';
 import Icon from '../Icon';
 
 const types = {
+  tiny: 'chip--tiny',
   slim: 'chip--slim',
   medium: 'chip--medium',
+  mediumBigText: 'chip--medium-big-text',
   big: 'chip--big',
 };