+++ src/common/components/PortfolioCard/index.jsx
@@ -26,23 +26,22 @@ const PortfolioCard = ({ name, iconBrandClass, industryIconName }) => {
           src={`/assets/img/portfolio/${name}.svg`}
         />
       </picture>
-      <div className="px-6 w-full h-full py-4 flex flex-col justify-center bg-white rounded-bl-[20px] rounded-br-[20px]">
+      <div className="px-6 w-full pt-4 pb-10 flex flex-col justify-center bg-white rounded-bl-[20px] rounded-br-[20px]">
         <div className="flex items-center justify-between">
           <Icon className={iconBrandClass} name={name} />
-          <Icon className="w-[25px] h-[26px] text-white" name={industryIconName} />
+          <Icon className="h-[28px] text-white" name={industryIconName} />
         </div>
         <p className="text-[13px] text-black xl:max-w-[440px] mt-8 line-clamp-2 xl:min-h-[39px]">
           {t(`portfolio.cards.${name}.text`)}
         </p>
-        <div className="mt-[48px] flex flex-col sm:flex-row space-y-1 sm:space-y-0 sm:space-x-4 items-center justify-center">
-          <p className="text-black text-[13px] font-medium">{t('portfolio.message')}</p>
-          <CustomNextLink to={`/${locale}/contact`}>
+        <div className="mt-[24px] flex flex-col sm:flex-row space-y-1 sm:space-y-0 sm:space-x-4 items-center justify-center">
+          <CustomNextLink to={`/${locale}/portfolio/${name}`}>
             <button
               aria-label={t('navbar.item4.ariaLabel')}
-              className="flex space-x-1 items-center justify-center text-xs font-bold text-blue sm:text-white bg-white sm:bg-blue w-fit sm:px-2 sm:py-1 rounded-[20px]"
+              className="flex space-x-2 items-center justify-center text-xs font-bold text-blue sm:text-white bg-white sm:bg-blue w-fit sm:px-4 sm:py-[6px] rounded-[20px]"
             >
               <p>{t(`portfolio.button`)}</p>
-              <Icon className="transform rotate-[270deg] w-[13px]" name="chevronDown" />
+              <Icon className="transform rotate-[270deg] w-4" name="chevronDown" />
             </button>
           </CustomNextLink>
         </div>