+++ CHANGELOG.md
@@ -5,7 +5,55 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Release 3.0.20 27/03/2024
+
+### Modified 19/04/2024
+
+- **Components**
+  - `PortfolioCard`: Component modified to redirect to client individual portfolio.
+
+### Added 19/04/2024
+
+- **Components**
+  - `FlipCardDescription`: A custom component to show de flipCard description with a text and shows an array of chips and a back button.
+
+- **Pages**
+  - `FinketPage`: Custom page for client protfolio.
+
+### Modified 12/04/2024
+
+- **Components**
+  - `FlipCard`: Component modified to recieve a TextButton.
+  - `TextButton`: Component modified to receive an Icon.
+
+### Added 12/04/2024
+
+- **Pages**
+  - `X28Page`: Custom page for client protfolio.
+
+### Modified 12/04/2024
+
+- **Modules**
+  - `WhatWeBuildSection`: Component modified to recieve children.
+
+### Added 11/04/2024
+
+- **Pages**
+  - `PowPage`: Custom page for client protfolio.
+
+### Added 04/04/2024
+
+- **Components**:
+  - `ApproachSection`: Custom module for client protfolio.
+  - `ChallengeSection`: Custom module for client protfolio.
+  - `MoreProjectSection`: Custom module for client protfolio.
+  - `PortfolioHeadingSection`: Custom module for client protfolio.
+  - `PortfolioHeaderSection`: Custom module for client protfolio.
+  - `WhatWeBuildSection`: Custom module for client protfolio.
+
+- **Pages**
+  - `SuraPage`: Custom page for client protfolio.
+
+## Release 3.1.0 27/03/2024
 
 ### Added 18/03/2024